web_sec_compliance/requirements.txt
streamlit==1.29.0
requests==2.31.0
dnspython==2.4.2
pandas==2.1.4